[refactor] dailyTopData를 새로 추가한 json파일로 교체 - 교체한 데이터는 필터를 설정하지 않아도 되기 때문에 필터와 관련된 코드를 삭제했다.
